Move mage stuff to magefiles directory
