--- internal/mage/grpc.go
@@ -1,12 +1,0 @@
-package mage
-
-import (
-	"fmt"
-)
-
-// RunBuf ...
-func RunBuf() error {
-	fmt.Printf("Running buf to generate GRPC code...\n")
-
-	return nil
-}
